+++ Demo/ChapterAccordionControl/ChapterAccordionControl.xaml.cs
@@ -1,0 +1,17 @@
+﻿// -----------------------------------------------------------------------------------------------------------------
+// <copyright file="ChapterAccordionControl.xaml.cs" company="my-libraries">
+//     Copyright (c) David Wendland. All rights reserved.
+// </copyright>
+// -----------------------------------------------------------------------------------------------------------------
+
+// ReSharper disable once CheckNamespace
+
+namespace Demo;
+
+public partial class ChapterAccordionControl
+{
+    public ChapterAccordionControl()
+    {
+        InitializeComponent();
+    }
+}