+++ Demo/ChapterButton/ChapterButtonControl.xaml.cs
@@ -4,14 +4,34 @@
 // </copyright>
 // -----------------------------------------------------------------------------------------------------------------
 
+using System.ComponentModel;
+
 // ReSharper disable once CheckNamespace
 
 namespace Demo;
 
-public partial class ChapterButtonControl
+public partial class ChapterButtonControl : INotifyPropertyChanged
 {
+    private int _iconSize;
+
     public ChapterButtonControl()
     {
         InitializeComponent();
+
+        IconSize = 21;
+
+        DataContext = this;
     }
+
+    public int IconSize
+    {
+        get => _iconSize;
+        set
+        {
+            _iconSize = value;
+            PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(nameof(IconSize)));
+        }
+    }
+
+    public event PropertyChangedEventHandler PropertyChanged;
 }